requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
rediscluster==2.1.3
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0
kafka-python==2.0.2
cryptography==41.0.7